Refactor and rename font conversion modules
The conversion modules 'ps_to_tt.py' and 'tt_to_ps.py' have been renamed to 'otf2ttf.py' and 'ttf2otf.py', respectively.
